Refactor in accordance to cursorless issue 1488 specs (#1)
Should complete the first bullet point of
https://github.com/cursorless-dev/cursorless/issues/1488 as mentioned in
the title. Things are a bit more hacky than I'd like, and in the future
I might rewrite a good chunk of the code in order to only use `git
clone`'d packages, but this is good enough for now. Obviously I can't
test anything publishing related, so this is as far as I can take
things.

Note that the parser packages listed in `package.json` are only the ones
present in
[vscode-parse-tree](https://github.com/cursorless-dev/vscode-parse-tree)
(plus gdscript since I have a pr open for that which this makes somewhat
irrelevant if I understand things correctly?); please lmk if the parser
packages that were unique to this project should be readded!

Finally, please note that a few parser packages have had their versions
upgraded. IIrc most of these were due to issues building older versions,
as nearly all of the packages which are `git clone`'d in `build.ts` at
runtime require special linking to build their `.wasm` file, which
thankfully is present in fairly recent versions of said repos.